requirements.txt
# Backend Requirements
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0

# AWS SDK
boto3==1.34.0
botocore==1.34.0

# HTTP Requests
requests==2.31.0

# Logging
coloredlogs==15.0.1
